refact: adjusting tests
