+++ app/src/main/java/com/picpay/desafio/android/base/BaseApplication.kt
@@ -1,10 +1,10 @@
 package com.picpay.desafio.android.base
 
 import android.app.Application
+import com.picpay.data.di.databaseModule
+import com.picpay.data.di.networkModule
 import com.picpay.desafio.android.BuildConfig
 import com.picpay.desafio.android.data.di.contactsModule
-import com.picpay.desafio.android.data.di.databaseModule
-import com.picpay.desafio.android.data.di.networkModule
 import com.picpay.desafio.android.presentation.di.agendaModule
 import com.picpay.desafio.android.utils.Constants.BASE_URL
 import org.koin.android.ext.koin.androidContext
