--- app/src/main/java/com/picpay/desafio/android/data/contacts/data/remote/PicPayService.kt
@@ -1,11 +1,0 @@
-package com.picpay.desafio.android.data.contacts.data.remote
-
-import com.picpay.desafio.android.data.contacts.data.remote.response.UserResponse
-import retrofit2.Response
-import retrofit2.http.GET
-
-interface PicPayService {
-
-    @GET("users")
-    suspend fun getUsers(): Response<List<UserResponse>>
-}
