--- app/src/main/java/com/picpay/desafio/android/data/contacts/data/remote/response/UserResponse.kt
@@ -1,8 +1,0 @@
-package com.picpay.desafio.android.data.contacts.data.remote.response
-
-data class UserResponse(
-    val id: Int,
-    val img: String?,
-    val name: String?,
-    val username: String?
-)
