+++ app/src/mock/java/com/picpay/desafio/android/data/di/ContactsModule.kt
@@ -1,6 +1,6 @@
 package com.picpay.desafio.android.data.di
 
-import com.picpay.desafio.android.data.contacts.data.repository.AgendaRepositoryMockImpl
+import com.picpay.data.contacts.data.repository.AgendaRepositoryMockImpl
 import com.picpay.domain.repository.AgendaRepository
 import org.koin.dsl.module
