+++ app/src/remote/java/com/picpay/desafio/android/data/di/ContactsModule.kt
@@ -1,8 +1,8 @@
 package com.picpay.desafio.android.data.di
 
-import com.picpay.desafio.android.data.contacts.data.local.dao.ContactsDao
-import com.picpay.desafio.android.data.contacts.data.remote.PicPayService
-import com.picpay.desafio.android.data.contacts.data.repository.AgendaRepositoryImpl
+import com.picpay.data.contacts.data.local.dao.ContactsDao
+import com.picpay.data.contacts.data.remote.PicPayService
+import com.picpay.data.contacts.data.repository.AgendaRepositoryImpl
 import com.picpay.domain.repository.AgendaRepository
 import org.koin.dsl.module
