+++ app/src/test/java/com/picpay/desafio/android/presentation/di/AgendaModuleTest.kt
@@ -2,10 +2,10 @@ package com.picpay.desafio.android.presentation.di
 
 import android.app.Application
 import com.picpay.desafio.android.BuildConfig
-import com.picpay.desafio.android.data.di.contactsModule
-import com.picpay.desafio.android.data.di.databaseModule
-import com.picpay.desafio.android.data.di.networkModule
+import com.picpay.data.di.databaseModule
+import com.picpay.data.di.networkModule
 import com.picpay.desafio.android.presentation.home.AgendaViewModel
+import com.picpay.desafio.android.data.di.contactsModule
 import com.picpay.desafio.android.utils.Constants
 import kotlinx.coroutines.test.TestCoroutineDispatcher
 import kotlinx.coroutines.test.TestCoroutineScope
