+++ data/build.gradle
@@ -1,44 +1,56 @@
+import com.picpay.buildSrc.Dependencies
+import com.picpay.buildSrc.Versions
+
 plugins {
     id 'com.android.library'
     id 'kotlin-android'
 }
 
+apply plugin: 'kotlin-kapt'
+
 android {
-    compileSdkVersion 30
-    buildToolsVersion "30.0.3"
+    compileSdkVersion Versions.compileSdk
 
     defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 30
-        versionCode 1
-        versionName "1.0"
+        minSdkVersion Versions.minSdk
+        targetSdkVersion Versions.targetSdk
+        versionCode Versions.versionCode
+        versionName Versions.versionName
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
     }
 
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = JavaVersion.VERSION_1_8.toString()
+        freeCompilerArgs = freeCompilerArgs + "-Xallow-result-return-type"
     }
 }
 
 dependencies {
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.2'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.3.0'
-    testImplementation 'junit:junit:4.+'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    api project(":domain")
+
+    Dependencies.Core.list.forEach { dependency ->
+        implementation dependency
+    }
+
+    Dependencies.Network.list.forEach { dependency ->
+        api dependency
+    }
+
+    Dependencies.DI.list.forEach { dependency ->
+        implementation dependency
+    }
+
+    kapt Dependencies.Core.roomCompiler
+
+    Dependencies.Test.testList.forEach { testDependency ->
+        testImplementation testDependency
+    }
 }
