+++ data/src/main/java/com/picpay/data/contacts/data/remote/response/UserResponse.kt
@@ -1,0 +1,8 @@
+package com.picpay.data.contacts.data.remote.response
+
+data class UserResponse(
+    val id: Int,
+    val img: String?,
+    val name: String?,
+    val username: String?
+)
