+++ data/src/main/java/com/picpay/data/contacts/data/utils/Constants.kt
@@ -1,0 +1,13 @@
+package com.picpay.data.contacts.data.utils
+
+object Constants {
+
+    const val DEFAULT_TIMEOUT = 60L
+    const val BASE_URL = "baseUrl"
+    const val GENERIC_NETWORK_ERROR = ""
+    const val DEFAULT_ERROR = ""
+    const val UNKNOWN_ERROR = ""
+
+    const val DATABASE_VERSION = 1
+    const val DATABASE_NAME = "agenda_database"
+}
