+++ data/src/test/java/com/picpay/data/utils/FakeModels.kt
@@ -1,19 +1,11 @@
-package com.picpay.desafio.android.utils
+package com.picpay.data.utils
 
 import com.picpay.data.contacts.data.local.entity.UserLocal
 import com.picpay.data.contacts.data.remote.response.UserResponse
-import com.picpay.domain.entity.ResponseHandler
 import com.picpay.domain.entity.User
-import com.picpay.domain.entity.ApiError
-import com.picpay.data.extensions.ErrorHandler
-import okhttp3.MediaType.Companion.toMediaTypeOrNull
-import okhttp3.ResponseBody.Companion.toResponseBody
-import retrofit2.HttpException
-import retrofit2.Response
 
 object FakeModels {
 
-    val FAKE_EMPTY_CONTACTS: List<UserResponse>? = null
     val FAKE_CONTACTS_RESPONSE = listOf(
         UserResponse(
             id = 1,
@@ -74,23 +66,4 @@ object FakeModels {
             username = "Empty username"
         )
     )
-    val GENERIC_ERROR: Response<UserResponse> = Response.error(
-        404,
-        "{\"message\":\"Api Generic Error\"}"
-            .toResponseBody("application/json".toMediaTypeOrNull())
-    )
-    val GENERIC_UNEXPECTED_ERROR: Response<UserResponse> = Response.error(
-        409,
-        "{\"xxxxxxxxx\":\"123 test\"}"
-            .toResponseBody("application/json".toMediaTypeOrNull())
-    )
-    val GENERIC_ERROR_RESPONSE = ResponseHandler.Error(
-        ApiError.HttpError(
-            code = 404,
-            message = Constants.GENERIC_NETWORK_ERROR,
-            exception = ErrorHandler().getErrorFromApi(HttpException(GENERIC_ERROR)).exception
-        )
-    )
-    val SUCCESS_CONTACTS_CALL = ResponseHandler.Success(FAKE_CONTACTS)
-    val EMPTY_CONTACTS_CALL = ResponseHandler.Empty
 }
