+++ domain/build.gradle
@@ -4,6 +4,7 @@ plugins {
     id 'java-library'
     id 'kotlin'
 }
+apply plugin: 'kotlin-kapt'
 
 java {
     sourceCompatibility = JavaVersion.VERSION_1_8
